feat(importance_sampling): implement geometric importance sampling for photon generation
Add `use_importance_sampling` configuration flag to enable photon direction biasing towards the first lens. When enabled, photons are generated uniformly within a cone covering the lens aperture, with their statistical weight adjusted accordingly to maintain unbiased estimates. This significantly increases simulation efficiency for extended sources.

- Add `ImportanceSamplingHelper` class to calculate optimal cone axis and opening angle based on lens geometry and source position
- Modify `PrimaryGeneratorAction` to override photon direction and weight when importance sampling is active
- Update `EventAction` and `SensitiveDetector` to accumulate weighted hits (`double` instead of `int`)
- Propagate configuration flag through `ActionInitialization` to both optimization and lens simulation executables
- Update all analysis scripts (`plot2D`, `plot3D`, `psf_extractor`, etc.) to handle weighted hit counts (`Double_t` branches)
- Extend `DetectorConstruction` to provide full lens parameters for cone calculation
- Update documentation (`README.md`) to reflect new weighted hit fields

+++ analysis/plot2D.cpp
@@ -150,7 +150,7 @@ int main(int argc, char** argv) {
   double x2_val;
   int config_id;
   int n_photons_val;
-  int n_hits_val;
+  double n_hits_val;
   char l1_id_buf[256];
   char l2_id_buf[256];
 
@@ -179,8 +179,8 @@ int main(int argc, char** argv) {
 
   for (int i = 0; i < tree_config->GetEntries(); ++i) {
     tree_config->GetEntry(i);
-    std::string s1 = tree_config->GetBranch("lens75_id") ? l1_id_buf : "default";
-    std::string s2 = tree_config->GetBranch("lens60_id") ? l2_id_buf : "default";
+    std::string s1        = tree_config->GetBranch("lens75_id") ? l1_id_buf : "default";
+    std::string s2        = tree_config->GetBranch("lens60_id") ? l2_id_buf : "default";
     config_map[config_id] = {x1_val, x2_val, s1, s2};
 
     bool found = false;
